refactor GPU comparison table to fetch initial data server-side and simplify component structure; update footer to use supabase client from lib; adjust scraping script for Vast.ai source URL
