+++ lib/hooks/useGPUData.js
@@ -4,13 +4,19 @@ import { useState, useEffect } from 'react';
 import { fetchGPUPrices } from '@/lib/utils/fetchGPUData';
 import { useFilter } from '@/lib/context/FilterContext';
 
-export function useGPUData({ ignoreContext = false } = {}) {
-  const [gpuData, setGpuData] = useState([]);
-  const [loading, setLoading] = useState(true);
+export function useGPUData({ ignoreContext = false, initialData = null } = {}) {
+  const [gpuData, setGpuData] = useState(initialData || []);
+  const [loading, setLoading] = useState(!initialData);
   const [error, setError] = useState(null);
   const { selectedProvider, selectedGPU } = useFilter();
 
   useEffect(() => {
+    if (initialData) {
+      // If we have initial data, only fetch when filters change
+      if (ignoreContext || (!selectedProvider && !selectedGPU)) {
+        return;
+      }
+    }
     fetchData();
   }, [ignoreContext ? null : selectedProvider, ignoreContext ? null : selectedGPU]);
